Home endpoint now is correct. Gives the user access to creating and edeting a home

+++ backend/src/main/java/org/example/backend/service/HomeService.java
@@ -11,6 +11,7 @@ import org.example.backend.repro.HomeRepro;
 import org.example.backend.service.mapper.HomeMapper;
 import org.example.backend.service.security.IdService;
 import org.example.backend.service.security.exception.HomeDoesNotExistException;
+import org.example.backend.service.security.exception.UserDoesNotHavePermissionException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -30,18 +31,20 @@ public class HomeService {
     }
 
     public List<HomeTableReturnDTO> getAllHomes(String userId) {
-        return homeRepro.findAll().stream()
+        return homeRepro.findHomesByMemberId(userId).stream()
                 .map(homeMapper::mapToHomeTableReturn).toList();
-
     }
 
     public HomeTableReturnDTO createNewHome(String userId, CreateHomeDTO createHomeDTO) {
         Home home = homeMapper.mapToHome(createHomeDTO).withId(idService.createNewId());
+        home.members().put(userId, Role.ADMIN);
         return homeMapper.mapToHomeTableReturn(homeRepro.save(home));
     }
 
     public HomeTableReturnDTO editHome(String userId,String id, EditHomeDTO editHomeDTO) throws HomeDoesNotExistException {
         Home home = homeRepro.findById(id).orElseThrow(() ->(new HomeDoesNotExistException("No Home with this ID found")));
+        //Chek if User has permission
+        checkForPermission(home,userId);
 
         //Changes Values if Change is provided by User
         if(editHomeDTO.name() != null){
@@ -50,12 +53,6 @@ public class HomeService {
 
         if(editHomeDTO.address() != null){
             home = home.withAddress(editHomeDTO.address());
-        }
-
-        if(editHomeDTO.associatedUsers() != null){
-            for(String associatedUser : editHomeDTO.associatedUsers()){
-                home.members().put(associatedUser, Role.MEMBER);
-            }
         }
 
         //Saves Home
@@ -67,11 +64,14 @@ public class HomeService {
 
     public void deleteHome(String userId,String id) {
         //TODO What happens to connect Task
+        Home home = homeRepro.findById(id).orElseThrow(() ->(new HomeDoesNotExistException("No Home with this ID found")));
+        //Chek if User has permission
+        checkForPermission(home,userId);
         homeRepro.deleteById(id);
     }
 
     public List<HomeListReturnDTO> getHomeNames(String userId) {
-        return homeRepro.findAll().stream()
+        return homeRepro.findHomesByMemberId(userId).stream()
                 .map(homeMapper::mapToHomeListReturn)
                 .toList();
     }
@@ -79,5 +79,11 @@ public class HomeService {
 
     public List<String> findHomeConnectedToUser(String userId) {
         return homeRepro.findHomesByMemberId(userId).stream().map(Home::id).toList();
+    }
+
+    private void checkForPermission(Home home, String userId){
+        if(!home.members().containsKey(userId)){
+            throw new UserDoesNotHavePermissionException("User Does Not Have Permission");
+        }
     }
 }
